pyproject.toml
[tool.poetry]
name = "projectsummarizer"
version = "0.1.0"
description = "A tool to summarize project files with tree view and extension analysis"
authors = ["e_mamul <eugene@pufin.id>"]
readme = "README.md"
packages = [{include = "projectsummarizer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
nbformat = "^5.10.4"
tiktoken = "^0.8.0"
anthropic = "^0.45.2"
python-dotenv = "^1.0.1"
prettytable = "^3.9.0"
anytree = "^2.12"
pathspec = "^0.12"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
